Remove filometro.models package and move modules to filometro package
